remove the `lru_cache` and `listify` from the extractors.
`lru_cache` seems to make more damage than greatness, since it needs to compute the hash of the arguments, making it slower than just parsing the text in case of cache miss.
Extractors now just yield captures, it is a concern of the caller to make them a list or keep just the generator.